updated the validation script
