+++ validation_script.py
@@ -1,42 +1,5 @@
 #%%
-import glob
-import os, random
-import torch
-import torch.nn as nn
-import pandas as pd
-from scipy.io import loadmat
-import matplotlib.pyplot as plt
-import numpy as np
-import argparse
-import seaborn as sns
-os.environ['PYTHONHASHSEED'] = 'python'
-from IPython.display import Markdown, display
-from sklearn.preprocessing import label_binarize
-from sklearn.preprocessing import MinMaxScaler
-
-
-import torch
-import torch.nn as nn
-import torch.optim as optim
-from torch.utils.data import DataLoader, Dataset
-import matplotlib.pyplot as plt
-from sklearn.model_selection import train_test_split
-from tqdm import tqdm
-
-
-#For Reproducibility
-def reproducibility_requirements(seed=42):
-    random.seed(seed)
-    os.environ['PYTHONHASHSEED'] = str(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-    #torch.use_deterministic_algorithms(True)
-    torch.cuda.manual_seed_all(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    #print("Set seed of", str(seed),"is done for Reproducibility")
-
-reproducibility_requirements()
+from cosfire_workflow_utils import *
 
 
 # Hyperparameters
@@ -49,6 +12,7 @@ parser.add_argument('--batch_size', type=int, default=32, help='Batch size')
 parser.add_argument('--epochs', type=int, default=800, help='Number of epochs')
 parser.add_argument('--alpha', type=int, default= 1e-5)
 parser.add_argument('--props', type=int, default=0.1)
+parser.add_argument('--margin', type=int, default=36)
 
 
 args = parser.parse_args()
@@ -60,117 +24,10 @@ epochs = args.epochs
 data_path = args.data_path
 alpha = args.alpha
 props = args.props
+margin = args.margin
 
 
-def SimplifiedTopMap(rB, qB, retrievalL, queryL, topk):
-  '''
-    rB - binary codes of the training set - reference set,
-    qB - binary codes of the query set,
-    retrievalL - labels of the training set - reference set, 
-    queryL - labels of the query set, and 
-    topk - the number of top retrieved results to consider.
-
-    rB = train_binary
-    qB = valid_binary
-    retrievalL = train_label
-    queryL = valid_label
-    topk = 100
-  '''
-  num_query = queryL.shape[0]
-  mAP = [0] * num_query
-  for i, query in enumerate(queryL):
-    rel = (np.dot(query, retrievalL.transpose()) > 0)*1 # relevant train label refs.
-    hamm = np.count_nonzero(qB[i] != rB, axis=1) #hamming distance
-    ind = np.argsort(hamm) #  Hamming distances in ascending order.
-    rel = rel[ind] #rel is reordered based on the sorted indices ind, so that it corresponds to the sorted Hamming distances.
-
-    top_rel = rel[:topk] #contains the relevance values for the top-k retrieved results
-    tsum = np.sum(top_rel) 
-
-    #skips the iteration if there are no relevant results.
-    if tsum == 0:
-        continue
-
-    pr_count = np.linspace(1, tsum, tsum) 
-    tindex = np.asarray(np.where(top_rel == 1)) + 1.0 #is the indices where top_rel is equal to 1 (i.e., the positions of relevant images)
-    pr = pr_count / tindex # precision
-    mAP_sub = np.mean(pr) # AP
-    mAP[i] = mAP_sub 
-      
-
-
-  return np.round(np.mean(mAP),4) *100 #mAP
-
-
-
-def mAP_values(df_training,df_testing, thresh = 0.5, percentile = True):
-    if percentile:
-        train_binary = np.array([((out >= np.percentile(out,thresh))*1).tolist()  for _, out in enumerate(df_training.predictions)])
-        valid_binary = np.array([((out >= np.percentile(out,thresh))*1).tolist()  for _, out in enumerate(df_testing.predictions)])
-    else:
-        train_binary = np.array([((out >= thresh) * 1).tolist() for _, out in enumerate(df_training.predictions)])
-        valid_binary = np.array([((out >= thresh) * 1).tolist() for _, out in enumerate(df_testing.predictions)])
-
-    train_label = label_binarize(df_training.label_code, classes=[0, 1, 2,3])
-    valid_label = label_binarize(df_testing.label_code, classes=[0,1, 2,3])
-
-    rB = train_binary
-    qB = valid_binary
-    retrievalL = train_label
-    queryL = valid_label
-    topk = 100
-    mAP = SimplifiedTopMap(rB, qB, retrievalL, queryL, topk)
-  
-    return np.round(mAP,4), train_binary, train_label, valid_binary, valid_label
-
-
-def load_data(path):
-       
-   # Load the MATLAB file
-   data = loadmat(path)
-   df0 = pd.DataFrame(data['COSFIREdescriptor']['training'][0][0][0][0][0])
-   df0['label'] = 'FRI'
-   df1 = pd.DataFrame(data['COSFIREdescriptor']['training'][0][0][0][0][1])
-   df1['label'] = 'FRII'
-   df2 = pd.DataFrame(data['COSFIREdescriptor']['training'][0][0][0][0][2])
-   df2['label'] = 'Bent'
-   df3 = pd.DataFrame(data['COSFIREdescriptor']['training'][0][0][0][0][3])
-   df3['label'] = 'Compact'
-   df_training = pd.concat([df0, df1, df2, df3], ignore_index=True)
-
-   df0 = pd.DataFrame(data['COSFIREdescriptor']['testing'][0][0][0][0][0])
-   df0['label'] = 'FRI'
-   df1 = pd.DataFrame(data['COSFIREdescriptor']['testing'][0][0][0][0][1])
-   df1['label'] = 'FRII'
-   df2 = pd.DataFrame(data['COSFIREdescriptor']['testing'][0][0][0][0][2])
-   df2['label'] = 'Bent'
-   df3 = pd.DataFrame(data['COSFIREdescriptor']['testing'][0][0][0][0][3])
-   df3['label'] = 'Compact'
-   df_testing = pd.concat([df0, df1, df2, df3], ignore_index=True)
-  
-   # Rename the columns:
-   column_names = ["descrip_" + str(i) for i in range(1, 201)] + ["label_code"]
-   df_training.columns = column_names
-   df_testing.columns = column_names
-
-   dic_labels = { 'Bent':2,
-                  'Compact':3,
-                     'FRI':0,
-                     'FRII':1
-               }
-   df_training['label_name'] = df_training['label_code'].map(dic_labels)
-   df_testing['label_name'] = df_testing['label_code'].map(dic_labels)
-
-
-   df_training_new = pd.concat([df_training,df_testing], ignore_index=True)
-
-   train_label_code = df_training['label_name']
-   valid_label_code = df_testing['label_name']
-
-   df_training.drop('label_code', axis=1, inplace=True)
-   df_testing.drop('label_code', axis=1, inplace=True)
-
-   return df_training, df_testing, train_label_code, valid_label_code, df_training_new
+#model
 
 class CosfireNet(nn.Module):
     def __init__(self, input_size, output_size):
@@ -189,6 +46,8 @@ class CosfireNet(nn.Module):
     def forward(self, x):
         return self.hd(x)
 
+
+# Data
 class CosfireDataset(Dataset):
     def __init__(self, dataframe):
         self.data = torch.tensor(dataframe.iloc[:, :-1].values, dtype=torch.float32)
@@ -201,10 +60,12 @@ class CosfireDataset(Dataset):
         return self.data[idx], self.labels[idx]
 
 
-def DSHLoss(u, y, alpha):
+# loss function
+
+def DSHLoss(u, y, alpha, margin):
     #alpha = 1e-5  # m > 0 is a margin. The margin defines a radius around X1,X2. Dissimilar pairs contribute to the loss
     # function only if their distance is within this radius
-    m = 2 * output_size  # Initialize U and Y with the current batch's embeddings and labels
+    m = 2 * margin  # Initialize U and Y with the current batch's embeddings and labels
     y = y.int()
     # Create a duplicate y_label
     y = y.unsqueeze(1).expand(len(y),len(y))
@@ -226,8 +87,8 @@ def DSHLoss(u, y, alpha):
 
 
 # Grid search parameters
-learning_rate_values = [0.1, 0.01, 0.001, 0.0001]
-alphas = [1e-3,1e-4,1e-5]
+learning_rate_values = [0.01]
+alphas = [1e-5]
 
 
 def run():
@@ -235,7 +96,7 @@ def run():
         for alpha in alphas:
 
             path = args.data_path
-            df_training, df_testing, train_label_code, valid_label_code, _ = load_data(path)
+            df_training, df_testing, train_label_code, valid_label_code, _ = get_data(path)
             
             train_df, valid_df = train_test_split(df_training, test_size=props, random_state=42)
             train_df.rename(columns = {'label_name': 'label_code'}, inplace = True)
@@ -270,7 +131,7 @@ def run():
                 for _, (inputs, labels) in enumerate(train_dataloader):
                     optimizer.zero_grad()
                     u = model(inputs)
-                    loss = DSHLoss(u = u, y=labels, alpha = alpha)
+                    loss = DSHLoss(u = u, y=labels, alpha = alpha, margin = margin)
                     loss.backward()
                     optimizer.step()
                     total_train_loss += loss.item()
@@ -287,7 +148,7 @@ def run():
                 with torch.no_grad():
                     for val_inputs, val_labels in val_dataloader:
                         val_outputs = model(val_inputs)
-                        val_loss = DSHLoss(u = val_outputs, y=val_labels, alpha = alpha)
+                        val_loss = DSHLoss(u = val_outputs, y=val_labels, alpha = alpha, margin = margin)
                         total_val_loss += val_loss.item()
 
                 # Calculate average validation loss
@@ -386,15 +247,16 @@ def run():
                 'mAP': maP,
                 'alpha': alpha,
                 'model_type' : 'model_v5',
-                "props": props
+                "props": props,
+                "margin": margin
                
                 
             }
             results_df = pd.DataFrame([results])
             
-            df = pd.read_csv("results.csv")
+            df = pd.read_csv("model_selection - validation.csv")
             results_df = pd.concat([df, results_df], ignore_index=True)
-            results_df.to_csv('results.csv', index=False)
+            results_df.to_csv('model_selection - validation.csv', index=False)
 
 
 if __name__ == '__main__' :
